routes.js moved to /api/index.js + /vercel.json added to allow serverless deployment on Vercel
